Implement Hash Join for O(N+M) performance with proper error handling
- Replaced nested-loop join with hash join in `relvar-core/src/algebra/join.rs`
- Optimized build side selection for minimal memory usage
- Updated `join` signature to return `Result<Relation, DatabaseError>`
- Replaced `.expect()` and silent error swallowing with error propagation
- Created `relvar-core/src/error.rs` to break circular dependencies
- Fixed call sites in benchmarks, tests, and examples
- Verified with benchmarks (4x speedup on 50k x 1k join)
