+++ relvar-core/src/lib.rs
@@ -16,6 +16,7 @@
 pub mod algebra;
 pub mod constraints;
 pub mod database;
+pub mod error;
 pub mod storage_engine;
 pub mod types;
 pub mod values;
